requirements.txt
pygame>=2.1.2
pyserial>=3.5
serial-interface>=2.1.0